add basic CRUD api
