+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/user/UserController.java
@@ -1,8 +1,10 @@
 package com.kbtg.bootcamp.posttest.user;
 
-import com.kbtg.bootcamp.posttest.userLottery.UserLotteryRequest;
 import com.kbtg.bootcamp.posttest.lottery.Lottery;
 import com.kbtg.bootcamp.posttest.lottery.LotteryService;
+import com.kbtg.bootcamp.posttest.userLottery.UserLotteryRequest;
+import com.kbtg.bootcamp.posttest.userLottery.UserLotteryResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -25,14 +27,15 @@ public class UserController {
     public List<User> getAllUser(){
         return userService.getAllUsers();
     }
+
+
     @GetMapping("/users/{id}")
     public User getUserById(@PathVariable("id") String id) {
-        // Implementation to retrieve user by ID
         return (User) userService.getUserById(id);
     }
 
-
     @PostMapping("/users")
+    @ApiResponse(responseCode = "201", description = "User Created")
     public User createUser(@RequestBody UserRequest request) {
         return userService.createUser(request);
     }
@@ -47,9 +50,19 @@ public class UserController {
         userService.buyLottery(userLotteryRequest.userId(), userLotteryRequest.lotteryId());
     }
 
+    @GetMapping("/users/{userId}/lotteries")
+    public UserLotteryResponse getUserLotteries(@PathVariable("userId") String userId) {
+        List<String> tickets = userService.getUserLotteryTickets(userId);
+        int count = tickets.size();
+        int cost = count * 80;
 
-    @DeleteMapping("/{userId}/lotteries/sell")
-    public void sellAllLotteries(@PathVariable("userId") String userId) {
+        return new UserLotteryResponse(tickets, count, cost);
+    }
 
+    @DeleteMapping("/users/{userId}/lotteries/sell")
+    public User sellAllLotteries(@PathVariable String userId) {
+        userService.deleteLotteries(userId);
+
+        return (User) userService.getUserById(userId);
     }
 }
